+++ src/main/java/com/example/CropMonitoringSystem/controller/FieldController.java
@@ -31,7 +31,6 @@ public class FieldController {
 
             Map<String, String> responseBody = new HashMap<>();
             responseBody.put("fieldId", fieldId);
-
             return new ResponseEntity<>(responseBody,HttpStatus.CREATED);
         } catch (DataPersistException e) {
             e.printStackTrace();
@@ -69,8 +68,12 @@ public class FieldController {
     @PreAuthorize("hasRole('MANAGER') or hasRole('SCIENTIST')")
     public ResponseEntity<Void> updateField(@PathVariable("fieldId") String fieldId,@RequestBody FieldDto fieldDto) {
         try {
-            fieldService.updateField(fieldId, fieldDto);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            if (!fieldId.matches(Regex.FIELD_ID)) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            } else {
+                fieldService.updateField(fieldId, fieldDto);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }
         } catch (DataPersistException e) {
             e.printStackTrace();
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
@@ -84,8 +87,12 @@ public class FieldController {
     @PreAuthorize("hasRole('MANAGER') or hasRole('SCIENTIST')")
     public ResponseEntity<Void> deleteField(@PathVariable("fieldId") String fieldId) {
         try {
-            fieldService.deleteField(fieldId);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            if (!fieldId.matches(Regex.FIELD_ID)) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            } else {
+                fieldService.deleteField(fieldId);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
@@ -99,6 +106,9 @@ public class FieldController {
 
     @GetMapping(value = "/{fieldId}", produces = MediaType.APPLICATION_JSON_VALUE)
     public FieldDto getFieldById(@PathVariable("fieldId") String fieldId) {
+        if (!fieldId.matches(Regex.FIELD_ID)) {
+            return null;
+        }
         return fieldService.getSelectedField(fieldId);
     }
 }
